refactor(platform): move enrichment dedup-state parsing into dedup subpackage (#756)
Continues the pkg/platform decomposition. The three pure functions that parse
persisted enrichment_dedup session state into middleware.SentTableEntry
(parseDedupState/parseDedupEntry/parseEntryFromMap) move out of platform.go
into pkg/platform/dedup as dedup.ParseState plus two unexported helpers — the
subpackage already documented as the home for dedup code split out of platform
to hold its size budget. loadPersistedEnrichmentState now calls dedup.ParseState;
the flush/load Platform methods stay. TestParseDedupState relocates alongside
as TestParseState. Verbatim move (renames only); make verify green.
